Point Click finished
Don't mess with this, it doesn't interfere with any code and can still run on robots. I finally got it to when you click on something, it leaves a mark there, this is a long wa from being completed, but it's out there so others can imporve later or give suggestions.