Fixed some path action stuff
read title, also we need to have a talk about whats going on with motion profiling... Are we giving the code a set of instructions and asking it to calculate a location based on those instructions? or are we giving it a location and having it generate a set of instructions based on that locations?